Picture Uploader endpoint.
Modified the FileAttachment model to add the id of the trip in the path.
This will probably break the current implementation of file attachments
on the web admin site, since the trip model is not instantiated on the
field attachment because of the use of “ContentType”. Any thoughts on
this would be helpful.