workaround
Given that the partner staff officer that can sign the agreement is
pre-populated, the “Partner Authorized Officers” is only present in
order to add more partner staff members as focal points for the
agreement. This ability is only available after the Agreement is saved.
Therefore the section will not be displayed unless the agreement has
been saved.