small tweaks
I’ve realized that there is no way of knowing the intervention total
cash transferred / CY or / CP based on the FC’s that we’re getting.
We’re not getting the FC dates only the FR dates. In theory you can
have FRs that are made outside of the CP or CY. Removing
actual_cash_transferred and total_cash_transferred until FC dates come
in from Insight