Replaced 3rd party library django-hstore with native Django functionality (hstore has been native since django 1.8)
Removing this library removes an impediment to upgrading to Django 1.10.
django-hstore isn't compatible with Django 1.10. The author is aware of this
but has no more time to maintain the module and is looking for a new
maintainer.
https://github.com/djangonauts/django-hstore/issues/154
https://github.com/djangonauts/django-hstore/issues/161

Also fixed all flake8 complaints in the file I touched (partners/models.py).