Django 1.10 prep -- no functional change. Remove references to django.conf.urls.patterns(), remove pattern prefix, refactor namespace handling.
patterns (as in `urlpatterns = patterns(..)`) was deprecated in Django 1.8 and
doesn't exist in Django 1.10. As part of the same change, the patterns prefix
(typically a blank string) was also deprecated.

In Django 1.10, everything in a list of URL patterns must be the result of a
call to url(). If you include a string in that list you'll get this unhelpful
error --
`AttributeError: 'str' object has no attribute 'regex'`
https://docs.djangoproject.com/en/1.11/releases/1.8/#django-conf-urls-patterns

In a related change, "referencing views as strings in the url() function is
deprecated" as of Django 1.8. There were a few instances of this in
`EquiTrack/urls.py`.
https://docs.djangoproject.com/en/1.11/releases/1.8/#passing-a-string-as-view-to-url

Removing the use of pattern() caused Django to begin raising this exception --
`ImproperlyConfigured: Cannot override the namespace for a dynamic module that provides a namespace`

That was as a result of this ticket which made important changes to URL
namespace declaration --
https://code.djangoproject.com/ticket/21927

To fix that I had to change how namespaces are specified. This mostly
affected `t2f/urls.py`. For a typical example, follow the diff in this commit
for `action_points_patterns` in `t2f/urls.py`.

Also fixed a typo in `t2f/urls.py`; changed this --
travel_pattens
to this --
travel_patterns