Change ints using long int syntax to regular ints for Python 3 compatibility.
None of these ints were actually long (the largest was 256), so no functional
change.

Django (correctly) does not regard this as a model change that affects
migrations, so 'makemigrations' reports 'no changes detected'. The Postgres
tables are defined the same way regardless of whether these values are defined
as int or long in Python.