Make test less rigorous by dropping test for matching filename of uploaded file.
Django reserves the right to change the name of uploaded files to avoid
name conflicts. At present it just adds a random 7-character string to the
existing filename if there's a conflict; the code is here --
https://github.com/django/django/blob/1.9.13/django/core/files/storage.py#L84

Rather than write a test that depends on the exact behavior of Django's
internals, I'm just removing filename from the test.