Refactor planned visits on PartnerOrganization
Before it appeared that if provided with a planned visit
we would exclude that provided planned visit from the query, and then
add it to the summed result from the query.

This had the side affect of including provided planned visits that
had years for non current years.

It seems simpler to just query all the planned visits for the current year
and sum their programmatic value.