Many things aren't lists in python 3
In Python 3, many things that used to return lists now return iterators.
Examples: map, filter, zip, dict.keys().

This becomes important because we can't do the same operations on iterators
as we can on lists - we can't index into them, take their length,
compare them to a list, treat them as a boolean, or iterate over them
more than once.

Additionally, dict.keys() doesn't necessarily return the keys in the same
order in Python 3 as it did in Python 2, so in tests it's safest to
sort the result and compare to a previously-sorted result.