Fix test that dirties the database, breaking --keepdb
monitoring.tests.test_checks.TestCheckDB creates user objects,
but was subclassing unittest.TestCase. Change to BaseTenantTestCase.

The rest of the changes are to other tests that were using
unittest.TestCase. Changing them to django.tests.SimpleTestCase
will result in the tests failing if anything in them changes the
database, so we'll know they need to be changed to BaseTenantTestCase.